new cache api
this patch re-introduce the cache api from barrel.

Instead of sharing one cache from the private environment we are exposing
the cache api from rocksdb as a nif resource. In the process we remove
the private cache which was keeping the memory too long.

this add the following functions:

-  new_lru_cache/1,
-  new_clock_cache/1,
-  get_usage/1,
-  get_pinned_usage/1,
-  set_capacity/2,
-  get_capacity/1
-  release_cache/1

note:

The cache is a shared_ptr which should survive
even if the nif resource is gone. At some point we
may want to pass the cache directly to the database
and keep the resource around.