fix "aclocal 1.14 is missing" error
snappy configure is okay, the next step is make but make exist with
error it said it cannot find aclocal. This is because we don’t have some
autoconf(.ac) and automake(.am) files. We can touch these which should
fix the error without having to regenerate the configure files.