make tests more coherent.
approximate_memtable_stats was testing if the returned count >= 0 but in
the mean time if the size >= 6000. These seemed strange for me, so i'm
now testing if Size >= 0 also since apparently from time to time the
returned count is 0 and  (logically?) the size is 0 too.

Similarely i am now testing iof Size > 0 in aproximate_sizes_test since
it testing the same stuff.

Note: i think we observe this behaviour because we afaik accessing to a
a pointer. While the access is thread safe, the content of it is not.
(hence imo why it's called "approximate") . Refactoring the Db resource
object would improve the return.